requirements.txt
streamlit
Pillow
bcrypt